# api/oas-doc.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Review API
  description: >-
    This is an OpenAPI Specification for Fafago Reviews API 
  contact:
    name: Review API for Fafago Service
    email: fafago@gmail.com
    url: 'https://github.com/fafagorg/reviews'
  license:
    name: MIT
    url: 'http://opensource.org/licenses/MIT'
paths:
...
  /api/v1/reviews:
    x-swagger-router-controller: reviewController
    get:
      description: Returns all reviews
      operationId: getReviews
      tags:
        - Reviews
      parameters:
        - name: Authorization
          in: header
          description: Token needed for authentication
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Review List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/reviewResponse'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
    post:
      description: Creates a new review
      operationId: addReview
      tags:
        - Reviews
      parameters:
        - name: Authorization
          in: header
          description: Token needed for authentication
          schema:
            type: string
          required: true
      responses:
        '201':
          description: Review created
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
      requestBody:
        x-name: review
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/review'
        description: Review to be created
        required: true
  '/api/v1/reviews/{id}':
    x-swagger-router-controller: reviewController
    get:
      description: Returns a review by id
      operationId: findReviewByid
      tags:
        - Reviews
      parameters:
        - name: id
          in: path
          description: id of review to fetch
          required: true
          schema:
            type: string
        - name: Authorization
          in: header
          description: Token needed for authentication
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Review returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/reviewResponse'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
    delete:
      description: deletes a single review based on the id
      operationId: deleteReview
      tags:
        - Reviews
      parameters:
        - name: id
          in: path
          description: ID of review to delete
          required: true
          schema:
            type: string
        - name: Authorization
          in: header
          description: Token needed for authentication
          schema:
            type: string
          required: true
      responses:
        '204':
          description: Review deleted
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
    put:
      description: Update a review based on its id
      operationId: updateReview
      tags:
        - Reviews
      parameters:
        - name: id
          in: path
          description: ID of review to put
          required: true
          schema:
            type: string
        - name: Authorization
          in: header
          description: Token needed for authentication
          schema:
            type: string
          required: true
      responses:
        '204':
          description: Review updated
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
      requestBody:
        x-name: review
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/review'
        description: Review to be updated
        required: true
  '/api/v1/reviews/client/{clientId}':
    x-swagger-router-controller: reviewController
    get:
      description: Returns all the reviews of the given client id
      operationId: findReviewsByClientId
      tags:
        - Reviews
      parameters:
        - name: clientId
          in: path
          description: id of the reviewed client
          required: true
          schema:
            type: string
        - name: Authorization
          in: header
          description: Token needed for authentication
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Reviews returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/reviewResponse'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
    delete:
      description: Deletes all the reviews of the given client id
      operationId: deleteReviewsByClientId
      tags:
        - Reviews
      parameters:
        - name: clientId
          in: path
          description: ID of the reviewed client
          required: true
          schema:
            type: string
        - name: Authorization
          in: header
          description: Token needed for authentication
          schema:
            type: string
          required: true
      responses:
        '204':
          description: Reviews deleted
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
  '/api/v1/reviews/product/{productId}':
    x-swagger-router-controller: reviewController
    get:
      description: Returns all the reviews of the given product id
      operationId: findReviewsByProductId
      tags:
        - Reviews
      parameters:
        - name: productId
          in: path
          description: id of the reviewed product
          required: true
          schema:
            type: string
        - name: Authorization
          in: header
          description: Token needed for authentication
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Reviews returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/reviewResponse'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
    delete:
      description: Deletes all the reviews of the given product id
      operationId: deleteReviewsByProductId
      tags:
        - Reviews
      parameters:
        - name: productId
          in: path
          description: ID of the reviewed product
          required: true
          schema:
            type: string
        - name: Authorization
          in: header
          description: Token needed for authentication
          schema:
            type: string
          required: true
      responses:
        '204':
          description: Reviews deleted
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
  '/api/v1/review/{id}/comments':
    x-swagger-router-controller: reviewController
    get:
      description: Returns all the comments of the review with the given id
      operationId: findReviewCommentsById
      tags:
        - Reviews
      parameters:
        - name: id
          in: path
          description: id of review to fetch comments
          required: true
          schema:
            type: string
        - name: Authorization
          in: header
          description: Token needed for authentication
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Comments returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/commentResponse'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
    post:
      description: Creates a new comment for the given review
      operationId: addCommentToReview
      tags:
        - Reviews
      parameters:
        - name: id
          in: path
          description: id of review to add the comment
          required: true
          schema:
            type: string
        - name: Authorization
          in: header
          description: Token needed for authentication
          schema:
            type: string
          required: true
      responses:
        '201':
          description: Comment added
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
      requestBody:
        x-name: comment
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/comment'
        description: Comment to be added
        required: true
  '/api/v1/review/{reviewId}/comment/{commentId}':
    x-swagger-router-controller: reviewController
    get:
      description: Returns the comment with the given id in the given review
      operationId: findReviewSingleComment
      tags:
        - Reviews
      parameters:
        - name: reviewId
          in: path
          description: id of the review
          required: true
          schema:
            type: string
        - name: commentId
          in: path
          description: id of the comment
          required: true
          schema:
            type: string
        - name: Authorization
          in: header
          description: Token needed for authentication
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Comment returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/commentResponse'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
    delete:
      description: Deletes the given comment from the given review
      operationId: deleteCommentFromReview
      tags:
        - Reviews
      parameters:
        - name: reviewId
          in: path
          description: Id of the review
          required: true
          schema:
            type: string
        - name: commentId
          in: path
          description: Id of the comment
          required: true
          schema:
            type: string
        - name: Authorization
          in: header
          description: Token needed for authentication
          schema:
            type: string
          required: true
      responses:
        '204':
          description: Comment deleted
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
    put:
      description: Updates the comment with the given id from the given review
      operationId: updateCommentFromReview
      tags:
        - Reviews
      parameters:
        - name: reviewId
          in: path
          description: Id of the review
          required: true
          schema:
            type: string
        - name: commentId
          in: path
          description: Id of the comment
          required: true
          schema:
            type: string
        - name: Authorization
          in: header
          description: Token needed for authentication
          schema:
            type: string
          required: true
      responses:
        '204':
          description: Comment updated
        '404':
          description: Comment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
      requestBody:
        x-name: comment
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/comment'
        description: Comment to be updated
        required: true
components:
  schemas:
    errorModel:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
    review:
      type: object
      properties:
        reviewedClientId:
          type: string
          minLength: 1
        reviewedProductId:
          type: string
          minLength: 1
        reviewerClientId:
          type: string
          minLength: 1
        title:
          type: string
          minLength: 1
        score:
          type: number
        description:
          type: string
          minLength: 1
        images: {}
        comments:
          type: array
          uniqueItems: false
          minItems: 0
          items:
            required:
              - clientId
            properties:
              clientId:
                type: string
                minLength: 1
              body:
                type: string
                minLength: 1
              date:
                type: string
                minLength: 1
      required:
        - reviewedClientId
        - reviewedProductId
        - reviewerClientId
        - title
        - score
        - description
...
    reviewResponse:
      type: object
      properties:
        id:
          type: string
          minLength: 1
        reviewedClientId:
          type: string
          minLength: 1
        reviewedProductId:
          type: string
          minLength: 1
        reviewerClientId:
          type: string
          minLength: 1
        title:
          type: string
          minLength: 1
        score:
          type: number
        description:
          type: string
          minLength: 1
        dateCreated:
          type: string
          minLength: 1
        images: {}
        comments:
          type: array
          uniqueItems: false
          minItems: 0
          items:
            required:
              - clientId
            properties:
              id:
                type: string
                minLength: 1
              clientId:
                type: string
                minLength: 1
              body:
                type: string
                minLength: 1
              date:
                type: string
                minLength: 1
      required:
        - id
        - dateCreated
        - reviewedClientId
        - reviewedProductId
        - reviewerClientId
        - title
        - score
        - description
        - comments
    comment:
      type: object
      properties:
        clientId:
          type: string
          minLength: 1
        body:
          type: string
          minLength: 1
      required:
        - clientId
        - body
...
    commentResponse:
      type: object
      properties:
        id:
          type: string
          minLength: 1
        clientId:
          type: string
          minLength: 1
        body:
          type: string
          minLength: 1
        date:
          type: string
          minLength: 1
      required:
        - id
        - clientId
        - body
        - date